pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-reliability-triage-env"
version = "0.1.0"
description = "Reliability incident triage benchmark environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "pydantic>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "reliability_triage_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["reliability_triage_env", "reliability_triage_env.server"]
package-dir = { "reliability_triage_env" = ".", "reliability_triage_env.server" = "server" }

[tool.pytest.ini_options]
testpaths = ["tests"]
